+++ graph/graph_test.go
@@ -710,7 +710,7 @@ func (c *tracedCounter) Forward(inputs ...*autograd.Variable) *autograd.Variable
 	return inputs[0] // pass through
 }
 func (c *tracedCounter) Parameters() []*nn.Parameter { return nil }
-func (c *tracedCounter) Reset(batchSize int64) {
+func (c *tracedCounter) Reset(batchSize int64, device tensor.Device) {
 	c.count = 0
 	c.resetCt++
 	c.lastBatch = batchSize